Consolidate into more test files
